Post 모델 수정 및 migrate 다시 하기
