+++ django/week01/comments/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from .models import Comments
 
 # Register your models here.
+admin.site.register(Comments)
