+++ django/week01/posts/admin.py
@@ -1,7 +1,5 @@
 from django.contrib import admin
 from .models import Post
-from .models import Comments
 
 # Register your models here.
 admin.site.register(Post)
-admin.site.register(Comments)
